+++ src/CDA.Consumer/GenericObjectReuseSample.cs
@@ -511,9 +511,11 @@ namespace Nehta.VendorLibrary.CDA.Consumer
                                                                     code);
 
               entitlement1.Type = EntitlementType.MedicarePharmacyApprovalNumber;
-              entitlement1.ValidityDuration = BaseCDAModel.CreateInterval("1", TimeUnitOfMeasure.Year);
+              entitlement1.ValidityDuration = BaseCDAModel.CreateInterval(
+                  new ISO8601DateTime(DateTime.Now, ISO8601DateTime.Precision.Day),
+                  new ISO8601DateTime(DateTime.Now.AddYears(1), ISO8601DateTime.Precision.Day));
 
-              person.Entitlements = new List<Entitlement> { entitlement1, entitlement1 };
+                person.Entitlements = new List<Entitlement> { entitlement1, entitlement1 };
 
               // Qualifications
               person.Qualifications = "M.B.B.S., F.R.A.C.S.";